Move tests to their own files, simplify some code
