+++ crypto/stark/src/tests/mod.rs
@@ -1,5 +1,8 @@
 pub mod air_tests;
+pub mod boundary_tests;
 pub mod bus_tests;
+pub mod fri_tests;
 pub mod proof_options_tests;
 pub mod prove_verify_roundtrip_tests;
+pub mod prover_tests;
 pub mod small_trace_tests;
